Rename workflow authoring modules to workflow define
Drop the separate 'authoring' product name in favor of workflow_define
modules that match the public defineWorkflow API. Fix Go joinWorkflowPath
for production scope helpers and update test references.
